Just fixed return type for IReceivingRadioi

+++ SPOTMonitor/src/org/sunspotworld/IReceivingRadio.java
@@ -16,7 +16,7 @@ public interface IReceivingRadio
 	 * Receives a datagram sent by the SendingRadio containing light data
 	 * @return light data
 	 */
-	public int receiveLight() throws IOException;
+	public double receiveLight() throws IOException;
 
 	/**
 	 * Receives a datagram sent by the SendingRadio containing heat data
